refactor: 디젝토리 구조 변경 (#38)
