+++ src/main/java/project/backend/business/post/implement/PostManager.java
@@ -3,7 +3,6 @@ package project.backend.business.post.implement;
 import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import project.backend.business.common.DateTimeManager;
 import project.backend.business.post.response.dto.PostDetailDto;
 import project.backend.business.tag.implement.TagManager;
 import project.backend.entity.post.Post;
